fix: Query cgroup memory data and fuse with /proc data for more reliable information.
The standard API for memory limits and usage under /proc are not
containerized, and return host data when used in a container. This is
a source of trouble for everything trying to configure itself based on
available memory, causing it to work with bogus data.

Inside containers the cgroup information rules instead.

Trying to detect when running inside a container or not looks to have
many approaches, which are either unreliable, or non-portable, or both.

The approach here is to query both sets of data, and then fuse them for
accuracy. As we ask for the available memory taking the min of cgroup and
proc data is the safe choice for this fusion.

It assumes that inside a container the cgroup data is likely less than
the proc data, due to the container getting limited. And outside the
container the cgroup data should not be less than proc.

(That said, in the local box, having 36GiB memory cgroup reports a
limit of nearly 8EiB (yes, Exa; Exact value is 8EiB-4KiB; As far as is
known the absolute limit for Linux virtual memory management).

As proc reports the 36GiB correctly the min is again the proper choice
for such systems where cgroup on the host is that way out of whack).

As for containers without limits set by docker or other runtime, cgroup
data seems to be that of the host. The min-fusion again gives the best
possible result, i.e. the host memory.

Note that getting the cgroup information is convoluted.

  - We have v1 and v2 controllers. Both can be mounted, some mixture,
    or neither.

  - The location of the mounts is arbitrary.

  - The process the library is part of is likely not managed by the
    root cgroup. The actual controlling group is needed.

  - Swap usage is not present if swap accounting is not active.

The gory details are noted in code comments.

Added tests.
Added clean up of the test's temp directory.
Added for most of the mock file setup for more readable tests.