Actually run Linux tests in CI
If you pass ",$SOMETHING" to the `-skipPackage` parameter of `ginkgo`,
`ginkgo` will skip tests in `$SOMETHING` suite, as well as those in
the current working directory. This means that the `gosigar` Linux
tests haven't been running in CI since ~2017.

This commit fixes that problem.

Additionally, I suspect that the `running_in_container` function only
works by coincidence and will emit false positives on a system that
doesn't have `cgroups` enabled, but it works in CI at the moment,
so I'm going to leave it alone.