Don't panic when cgroup controller multi-mounted
This commit removes the calls to 'panic' when more than one mount
point for either the cgroupv2 controller or the cgroupv1 memory
controller is found.

It turns out that it is perfectly legal to have a cgroup controller
mounted at multiple points in the filesystem. It is now known that
some Linux security software does exactly this.

[#180668200]

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>
Co-authored-by: Joseph Palermo <jpalermo@pivotal.io>